add update in todos for their own users

+++ src/app/api/todos/route.ts
@@ -1,55 +1,73 @@
+
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
-import { NextResponse } from 'next/server'
+import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
 
 export async function GET(request: Request) { 
+  
+  const { searchParams } = new URL(request.url)
+  const take = Number( searchParams.get('take') ?? '10');
+  const skip = Number( searchParams.get('skip') ?? '0');
 
-    const { searchParams } = new URL(request.url)
-    const take = Number(searchParams.get('take') ?? '10');
-    const skip = Number(searchParams.get('skip') ?? '0');
+  if ( isNaN(take) ){ 
+    return NextResponse.json({ message: 'Take tiene que ser un número' }, { status: 400 });
+  }
 
-    if( isNaN(take) ) {
-        return NextResponse.json({ message: 'Take tiene que ser un número'}, {status: 400});
-    }
-    if( isNaN(skip) ) {
-        return NextResponse.json({ message: 'Skip tiene que ser un número'}, {status: 400});
-    }
+  if ( isNaN(skip) ){ 
+    return NextResponse.json({ message: 'Skip tiene que ser un número' }, { status: 400 });
+  }
 
-    const todos = await prisma.todo.findMany({ take, skip });
+  const todos = await prisma.todo.findMany({
+    take: take,
+    skip: skip,
+  });
 
-    return NextResponse.json(todos);
+  return NextResponse.json( todos );
 }
 
+
+
 const postSchema = yup.object({
-    description: yup.string().required(),
-    complete: yup.boolean().optional().default(false),
-})
+  description: yup.string().required(),
+  complete: yup.boolean().optional().default(false),
+});
 
 export async function POST(request: Request) { 
 
-    try {
-        const { complete, description } = await postSchema.validate(await request.json());
+  const user = await getUserSessionServer();
 
-        const todo = await prisma.todo.create({data: {complete, description}});
+  if ( !user ) {
+    return NextResponse.json('No autorizado',{ status: 401 });
+  }
 
-        return NextResponse.json(todo);
 
-    } catch (error) {
+  try {
+    const { complete, description } = await postSchema.validate( await request.json() );
 
-        return NextResponse.json(error, {status: 400})
-
-    }
-
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
+  
+    
+    return NextResponse.json(todo);
+    
+  } catch (error) {
+    return NextResponse.json( error, { status: 400 } );
+  }
 }
-
 export async function DELETE(request: Request) { 
 
-    try {
-  
-      await prisma.todo.deleteMany({ where: { complete: true } });
-      return NextResponse.json('Borrados');
-      
-    } catch (error) {
-      return NextResponse.json( error, { status: 400 } );
-    }
+  const user = await getUserSessionServer();
+
+  if ( !user ) {
+    return NextResponse.json('No autorizado',{ status: 401 });
   }
+
+  try {
+
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
+    return NextResponse.json('Borrados');
+    
+  } catch (error) {
+    return NextResponse.json( error, { status: 400 } );
+  }
+}
